fix(autoware_qp_interface): fix bugprone-narrowing-conversions warnings (#922)
* fix(autoware_qp_interface): fix bugprone-narrowing-conversions warnings


* style(pre-commit): autofix

---------

+++ common/autoware_qp_interface/src/osqp_interface.cpp
@@ -329,7 +329,7 @@ void OSQPInterface::updateBounds(
 
 int OSQPInterface::getIterationNumber() const
 {
-  return work_->info->iter;
+  return static_cast<int>(work_->info->iter);
 }
 
 std::string OSQPInterface::getStatus() const
